+++ src/js/records/entities/asset.record.js
@@ -7,6 +7,7 @@ export class AssetRecord {
 
     this.code = record.code
     this.owner = record.owner
+    this.assetType = record.assetType
     this.preissuedAssetSigner = record.preissuedAssetSigner
 
     this.availableForIssuance = record.availableForIssuance