+++ src/vue/forms/AssetUpdateForm.vue
@@ -41,18 +41,6 @@
         </div>
       </div>
 
-      <div class="app__form-row asset-update-form__kyc-required-row">
-        <div class="app__form-field">
-          <tick-field
-            v-model="form.information.policies"
-            :disabled="formMixin.isDisabled"
-            :cb-value="ASSET_POLICIES.requiresKyc"
-          >
-            {{ 'asset-form.kyc-required-lbl' | globalize }}
-          </tick-field>
-        </div>
-      </div>
-
       <div class="app__form-row">
         <div class="app__form-field">
           <file-field